Initial commit
I got reminded that git is a good idea >.<

What's in:
- Loading
- Saving
- Looking up books
- Giving books to players
- Updating books
- Distributing the book updates to players

What's not in yet:
- Configurability
- On-first-login book giving (depends on config)
- Perhaps, maybe re-do the bruteforce updating of books.

Bruteforce updating of books:
Currently, books are updated with brute force.  That is, a book in
inventory is updated even if it hasn't changed since it was last
updated.  ALL books in your inventory are forcibly updated every
time you close an inventory.
This is close to lunacy, but it's the best I've gotten so far.