Library shelves implemented
The new action, -library, opens an inventory stocked with the books
in the library.  Currently limited to 54 slots because the client
starts rendering it oddly at that point.

Other changes:
- Implemented a -save action, useful when autosave is off.
- Added a missing permissions to plugin.yml (.action.prepend).
- Changed library storage to TreeMap rather than HashMap.
- Re-creating the FileConfiguration on each save/load to be sure
  it's a fresh one.

On the TreeMap thing...  I realize that TreeMap is way way slower
than HashMap by a factor of ... well, a lot.
The thing is, that a TreeMap will automagically alphabetize the
library, so I don't have to do that myself, and I can't think of
a case where a library would have so many thousand books that you
could actually notice the difference.